Peles Castle Added
Added content for the construction of Peles Castle.
Added minor event for the Romanian national academy.
Reordered balkfm.630 -> balkfm.640; Block 63 will be for Transylvania content.
Added a framework of variables, effect, and modifiers for monument construction. This will be the standard for future monuments.
Removed radical fraction check in eqfm.020.