Greek Revolts Framework
Added "Greek Separatism" with a main and sub journal and scripting framework.
The Greek revolts have been modified to use a combination of radical fractions and scripting to make the revolts feel less contrived.
Removed and integrated the Cretan Revolts journals into this framework.
Removed scripted buttons and status_desc from the Megali Idea in preparation.